requirements.txt
absl-py==2.3.1
aiofiles==25.1.0
attrs==25.4.0
bibtexparser==1.4.3
certifi==2025.11.12
charset-normalizer==3.4.4
chex==0.1.91
contourpy==1.3.3
cycler==0.12.1
dm-haiku==0.0.16
e3nn-jax==0.20.8
einops==0.8.1
equinox==0.13.2
etils==1.13.0
filelock==3.20.0
flax==0.12.2
fonttools==4.61.1
frozendict==2.4.7
fsspec==2025.12.0
humanize==4.15.0
idna==3.11
importlib_resources==6.5.2
jax==0.8.2
jax-cuda13-pjrt==0.8.2
jax-cuda13-plugin==0.8.2
jax-md==0.2.27
jaxlib==0.8.2
jaxtyping==0.3.4
Jinja2==3.1.6
jmp==0.0.4
joblib==1.5.3
jraph==0.0.6.dev0
kiwisolver==1.4.9
llvmlite==0.46.0
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.8
mdurl==0.1.2
ml_collections==1.1.0
ml_dtypes==0.5.4
monty==2025.3.3
mpmath==1.3.0
msgpack==1.1.2
narwhals==2.14.0
nest-asyncio==1.6.0
networkx==3.6.1
numba==0.63.1
numpy==2.3.5
nvidia-cublas==13.0.0.19
nvidia-cuda-crt==13.1.80
nvidia-cuda-cupti==13.0.48
nvidia-cuda-nvcc==13.1.80
nvidia-cuda-nvrtc==13.0.48
nvidia-cuda-runtime==13.0.48
nvidia-cudnn-cu13==9.13.0.50
nvidia-cufft==12.0.0.15
nvidia-cufile==1.15.0.42
nvidia-curand==10.4.0.35
nvidia-cusolver==12.0.3.29
nvidia-cusparse==12.6.2.49
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu13==2.27.7
nvidia-nvjitlink==13.0.39
nvidia-nvshmem-cu13==3.3.24
nvidia-nvtx==13.0.39
nvidia-nvvm==13.1.80
opt_einsum==3.4.0
optax==0.2.6
orbax-checkpoint==0.11.31
orjson==3.11.5
packaging==25.0
palettable==3.3.3
pandas==2.3.3
pillow==12.0.0
plotly==6.5.0
protobuf==6.33.2
psutil==7.2.0
Pygments==2.19.2
pymatgen==2025.10.7
pynndescent==0.6.0
pyparsing==3.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
rich==14.2.0
ruamel.yaml==0.18.17
ruamel.yaml.clib==0.2.15
scikit-learn==1.8.0
scipy==1.16.3
setuptools==70.2.0
simplejson==3.20.2
six==1.17.0
spglib==2.6.0
sympy==1.14.0
tabulate==0.9.0
tensorstore==0.1.80
threadpoolctl==3.6.0
toolz==1.1.0
torch==2.9.1+cu130
torchvision==0.24.1+cu130
tqdm==4.67.1
treescope==0.1.10
triton==3.5.1
typing_extensions==4.15.0
tzdata==2025.3
umap-learn==0.5.11
uncertainties==3.2.3
urllib3==2.6.2
wadler_lindig==0.1.7
zipp==3.23.0
